requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.5
django-allauth==0.41.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-js-asset==1.2.2
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
